pyproject.toml
[project]
name = "ai-pdf-reader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.12.1",
    "langchain-community>=0.3.22",
    "pypdf>=5.4.0",
]
